refactor: Home page codes in a separate folder
